see desc
refresh of GUI for the addition of starting game and starting voice ports
temporarily enabled checking the active game log with every lobby. This increases the over head on the bot slightly but reduces the likelihood of corrupting the data stream from the log being missed once.
If it does get messed, the entire work flow / cycle is missed for bot server restarts etc.