see desc
Added support for public IP changing while game server is running.
bot will detect this and restart the server with the new IP if there are no players connected.

Added using hon username and password to authenticate to master server

Added new method of launching game servers which is to use the configuration in the commandline. Moving away from using configuration file so that we can combine all the slave logs again in the future along with the use of manager.